Cleanup directory structure, uninstall unused dependencies
