+++ package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "anchor-js": "^4.2.2",
     "debounce": "^1.2.0",
-    "jquery": "^3.5.0",
     "preact": "^10.4.6",
     "redux-zero": "^5.1.7"
   },
@@ -27,7 +26,6 @@
     "execa": "^4.0.0",
     "fs-extra": "^9.0.0",
     "jest": "^26.1.0",
-    "liquid-loader": "^0.1.4",
     "listr": "^0.14.3",
     "prettier": "2.0.2",
     "semver": "^7.2.2",
