--- src_js/NodeManagerComponent.d.ts
@@ -1,3 +1,0 @@
-export default interface NodeManagerComponent {
-  init();
-}
