+++ src_js/components/Topbar.tsx
@@ -2,7 +2,6 @@ import { h } from 'preact';
 import { useRef, useLayoutEffect } from 'preact/hooks';
 import IconType from './common/IconType';
 import InlineButton from './common/InlineButton';
-import Utilities from '../Utilities';
 
 type PropsType = {
   sidebarShown: boolean;
